converted all the parsers to linters
