+++ inlineplz/interfaces/base.py
@@ -3,7 +3,7 @@ from __future__ import absolute_import
 from __future__ import unicode_literals
 
 
-class InterfaceBase():
+class InterfaceBase:
     """Abstract base class for review interfaces"""
 
     def start_review(self):
