+++ inlineplz/linters/__init__.py
@@ -15,12 +15,7 @@ import time
 import traceback
 
 from inlineplz.util.system import run_command, install_linter
-from inlineplz.registry import (
-    get_linters,
-    get_patterns,
-    register_pattern,
-    register_linter,
-)
+from inlineplz.registry import get_linters, get_patterns, register_pattern
 
 from identify import identify
 
@@ -31,10 +26,35 @@ try:
 except ImportError:
     from scandir import scandir, walk  # noqa
 
-from inlineplz import parsers
 from inlineplz import message
 from inlineplz.util import system
 
+# import your linter here
+from inlineplz.linters.ansiblelint import AnsibleLintParser  # NOQA
+from inlineplz.linters.bandit import BanditParser  # NOQA
+from inlineplz.linters.codenarc import CodenarcParser  # NOQA
+from inlineplz.linters.detectsecrets import DetectSecretsParser  # NOQA
+from inlineplz.linters.dockerfilelint import DockerfileLintParser  # NOQA
+from inlineplz.linters.eclint import ECLintParser  # NOQA
+from inlineplz.linters.eslint import ESLintParser  # NOQA
+from inlineplz.linters.gherkinlint import GherkinLintParser  # NOQA
+from inlineplz.linters.gometalinter import GometalinterParser  # NOQA
+from inlineplz.linters.htmlhint import HTMLHintParser  # NOQA
+from inlineplz.linters.jscs import JSCSParser  # NOQA
+from inlineplz.linters.jshint import JSHintParser  # NOQA
+from inlineplz.linters.jsonlint import JSONLintParser  # NOQA
+from inlineplz.linters.markdownlint import MarkdownLintParser  # NOQA
+from inlineplz.linters.megacheck import MegacheckParser  # NOQA
+from inlineplz.linters.pmd import PMDParser  # NOQA
+from inlineplz.linters.proselint import ProselintParser  # NOQA
+from inlineplz.linters.prospector import ProspectorParser  # NOQA
+from inlineplz.linters.rflint import RobotFrameworkLintParser  # NOQA
+from inlineplz.linters.rstlint import RSTLintParser  # NOQA
+from inlineplz.linters.shellcheck import ShellcheckParser  # NOQA
+from inlineplz.linters.spotbugsmaven import SpotbugsMavenParser  # NOQA
+from inlineplz.linters.stylint import StylintParser  # NOQA
+from inlineplz.linters.tflint import TFLintParser  # NOQA
+from inlineplz.linters.yamllint import YAMLLintParser  # NOQA
 
 HERE = os.path.dirname(__file__)
 
@@ -57,10 +77,6 @@ def vendored_path(path):
     return os.path.normpath(
         os.path.abspath(os.path.join(os.path.dirname(HERE), "bin", path))
     )
-
-
-GROOVY_PATH = vendored_path(os.path.join("groovy", "groovy-all-2.4.15.jar"))
-SLF4J_PATH = vendored_path(os.path.join("groovy", "slf4j-api-1.7.25.jar"))
 
 
 def register_patterns():
@@ -86,650 +102,8 @@ def register_patterns():
     register_pattern("yaml", ["*.yaml", "*.yml"])
 
 
-def register_linters():
-    register_linter(
-        "ansible-lint",
-        {
-            "install": [[sys.executable, "-m", "pip", "install", "-U", "ansible-lint"]],
-            "help": ["ansible-lint", "-h"],
-            "run": ["ansible-lint", "-p"],
-            "rundefault": ["ansible-lint", "-p", "-c", "{config_dir}/.ansible-lint"],
-            "dotfiles": [".ansible-lint"],
-            "parser": parsers.AnsibleLintParser,
-            "language": "ansible",
-            "autorun": True,
-            "run_per_file": True,
-        },
-    )
-    register_linter(
-        "bandit",
-        {
-            "install": [[sys.executable, "-m", "pip", "install", "-U", "bandit"]],
-            "help": ["bandit", "-h"],
-            "run": ["bandit", "-f", "json", "-iii", "-ll", "-r", "."],
-            "rundefault": [
-                "bandit",
-                "-f",
-                "json",
-                "-iii",
-                "-ll",
-                "-r",
-                ".",
-                "-c",
-                "{config_dir}/bandit.yaml",
-            ],
-            "dotfiles": ["bandit.yaml"],
-            "parser": parsers.BanditParser,
-            "language": "python",
-            "autorun": True,
-            "run_per_file": False,
-        },
-    )
-    register_linter(
-        "codenarc",
-        {
-            "install": [],
-            "help": [
-                "java",
-                "-classpath",
-                "{}{}{}{}{}{}{}{}{}".format(
-                    GROOVY_PATH,
-                    JAVA_SEP,
-                    vendored_path(os.path.join("codenarc", "CodeNarc-1.1.jar")),
-                    JAVA_SEP,
-                    SLF4J_PATH,
-                    JAVA_SEP,
-                    vendored_path("codenarc"),
-                    JAVA_SEP,
-                    ".",
-                ),
-                "org.codenarc.CodeNarc",
-                "-help",
-            ],
-            "run": [
-                "java",
-                "-classpath",
-                "{}{}{}{}{}{}{}{}{}".format(
-                    GROOVY_PATH,
-                    JAVA_SEP,
-                    vendored_path(os.path.join("codenarc", "CodeNarc-1.1.jar")),
-                    JAVA_SEP,
-                    SLF4J_PATH,
-                    JAVA_SEP,
-                    vendored_path("codenarc"),
-                    JAVA_SEP,
-                    ".",
-                ),
-                "org.codenarc.CodeNarc",
-                "-includes=**/*.groovy,**/Jenkinsfile,**/jenkinsfile,**/...groovy",
-                "-report=console",
-                "-rulesetfiles={}".format(os.path.join(os.getcwd(), "codenarc.xml")),
-            ],
-            "rundefault": [
-                "java",
-                "-classpath",
-                "{}{}{}{}{}{}{}{}{}".format(
-                    GROOVY_PATH,
-                    JAVA_SEP,
-                    vendored_path(os.path.join("codenarc", "CodeNarc-1.1.jar")),
-                    JAVA_SEP,
-                    SLF4J_PATH,
-                    JAVA_SEP,
-                    vendored_path("codenarc"),
-                    JAVA_SEP,
-                    ".",
-                ),
-                "org.codenarc.CodeNarc",
-                "-includes=**/*.groovy,**/Jenkinsfile,**/jenkinsfile,**/...groovy",
-                "-report=console",
-                "-rulesetfiles=codenarc.xml",
-            ],
-            "dotfiles": ["codenarc.xml"],
-            "parser": parsers.CodenarcParser,
-            "language": "groovy",
-            "autorun": True,
-            "run_per_file": False,
-        },
-    )
-    register_linter(
-        "detect-secrets",
-        {
-            # TODO: switch this to installing from pypi once they release my fix from https://github.com/Yelp/detect-secrets/pull/69
-            # "install": [[sys.executable, "-m", "pip", "install", "-U", "detect-secrets"]],
-            "install": [
-                [
-                    sys.executable,
-                    "-m",
-                    "pip",
-                    "install",
-                    "-U",
-                    "detect-secrets",
-                ]
-            ],
-            "help": ["detect-secrets", "-h"],
-            "run": ["detect-secrets", "scan", "--all-files"],
-            "rundefault": ["detect-secrets", "scan", "--all-files"],
-            "dotfiles": [],
-            "parser": parsers.DetectSecretsParser,
-            "language": "all",
-            "autorun": True,
-            "run_per_file": False,
-        },
-    )
-    register_linter(
-        "dockerfile_lint",
-        {
-            "install": [["npm", "install", "dockerfile_lint"]],
-            "help": [os.path.normpath("./node_modules/.bin/dockerfile_lint"), "-h"],
-            "run": [
-                os.path.normpath("./node_modules/.bin/dockerfile_lint"),
-                "-j",
-                "-f",
-            ],
-            "rundefault": [
-                os.path.normpath("./node_modules/.bin/dockerfile_lint"),
-                "-j",
-                "-f",
-            ],
-            "dotfiles": [],
-            "parser": parsers.DockerfileLintParser,
-            "language": "docker",
-            "autorun": True,
-            "run_per_file": True,
-        },
-    )
-    register_linter(
-        "eclint",
-        {
-            "install": [["npm", "install", "eclint"]],
-            "help": [os.path.normpath("./node_modules/.bin/eclint"), "-h"],
-            "run": [os.path.normpath("./node_modules/.bin/eclint"), "check"],
-            "rundefault": [os.path.normpath("./node_modules/.bin/eclint"), "check"],
-            "dotfiles": [".editorconfig"],
-            "parser": parsers.ECLintParser,
-            "language": "all",
-            "autorun": False,
-            "run_per_file": True,
-            "run_if_dotfile_in_root": False,
-        },
-    )
-    register_linter(
-        "eslint",
-        {
-            "install": [["npm", "install", "eslint"]],
-            "help": [os.path.normpath("./node_modules/.bin/eslint"), "-h"],
-            "run": [os.path.normpath("./node_modules/.bin/eslint"), ".", "-f", "unix"],
-            "rundefault": [
-                os.path.normpath("./node_modules/.bin/eslint"),
-                ".",
-                "-f",
-                "unix",
-                "-c",
-                "{config_dir}/.eslintrc.js",
-                "--ignore-path",
-                "{config_dir}/.eslintignore",
-            ],
-            "dotfiles": [
-                ".eslintrc.yml",
-                ".eslintrc.yaml",
-                ".eslintignore",
-                ".eslintrc",
-                ".eslintrc.js",
-                ".eslintrc.json",
-            ],
-            "parser": parsers.ESLintParser,
-            "language": "javascript",
-            "autorun": True,
-            "run_per_file": False,
-        },
-    )
-    register_linter(
-        "gherkin-lint",
-        {
-            "install": [["npm", "install", "gherkin-lint"]],
-            "help": [os.path.normpath("./node_modules/.bin/gherkin-lint"), "--help"],
-            "run": [
-                os.path.normpath("./node_modules/.bin/gherkin-lint"),
-                ".",
-                "-f",
-                "json",
-            ],
-            "rundefault": [
-                os.path.normpath("./node_modules/.bin/gherkin-lint"),
-                ".",
-                "-f",
-                "json",
-                "-c",
-                "{config_dir}/.gherkin-lintrc",
-            ],
-            "dotfiles": [".gherkin-lintrc"],
-            "parser": parsers.GherkinLintParser,
-            "language": "gherkin",
-            "autorun": True,
-            "run_per_file": False,
-        },
-    )
-    register_linter(
-        "gometalinter.v2",
-        {
-            "install": [
-                ["go", "get", "-u", "gopkg.in/alecthomas/gometalinter.v2"],
-                ["gometalinter.v2", "--install", "--update"],
-            ],
-            "help": ["gometalinter.v2", "--install", "--update"],
-            "run": [
-                "gometalinter.v2",
-                "--enable-gc",
-                "--deadline=300s",
-                "--aggregate",
-                "--vendor",
-                "--json",
-                "-s",
-                "node_modules",
-                "-s",
-                "src",
-                "./...",
-            ],
-            "rundefault": [
-                "gometalinter.v2",
-                "--enable-gc",
-                "--deadline=300s",
-                "--aggregate",
-                "--vendor",
-                "--disable=lll",
-                "--json",
-                "-s",
-                "node_modules",
-                "-s",
-                "src",
-                "--config={config_dir}/.gometalinter.json",
-                "./...",
-            ],
-            "dotfiles": [".gometalinter.json"],
-            "parser": parsers.GometalinterParser,
-            "language": "go",
-            "autorun": False,
-            "run_per_file": False,
-        },
-    )
-    register_linter(
-        "gometalinter",
-        {
-            "install": [
-                ["go", "get", "-u", "github.com/alecthomas/gometalinter"],
-                ["gometalinter", "--install", "--update"],
-            ],
-            "help": ["gometalinter", "--install", "--update"],
-            "run": [
-                "gometalinter",
-                "--enable-gc",
-                "--deadline=300s",
-                "--aggregate",
-                "--vendor",
-                "--json",
-                "-s",
-                "node_modules",
-                "-s",
-                "src",
-                "./...",
-            ],
-            "rundefault": [
-                "gometalinter",
-                "--enable-gc",
-                "--deadline=300s",
-                "--aggregate",
-                "--vendor",
-                "--disable=lll",
-                "--json",
-                "-s",
-                "node_modules",
-                "-s",
-                "src",
-                "--config={config_dir}/.gometalinter.json",
-                "./...",
-            ],
-            "dotfiles": [".gometalinter.json"],
-            "parser": parsers.GometalinterParser,
-            "language": "go",
-            "autorun": True,
-            "run_per_file": False,
-        },
-    )
-    register_linter(
-        "htmlhint",
-        {
-            "install": [["npm", "install", "htmlhint"]],
-            "help": [os.path.normpath("./node_modules/.bin/htmlhint"), "-h"],
-            "run": [os.path.normpath("./node_modules/.bin/htmlhint"), "--format=json"],
-            "rundefault": [
-                os.path.normpath("./node_modules/.bin/htmlhint"),
-                "--format=json",
-                "--config={config_dir}/.htmlhintrc",
-            ],
-            "dotfiles": [".htmlhintrc"],
-            "parser": parsers.HTMLHintParser,
-            "language": "html",
-            "autorun": True,
-            "run_per_file": False,
-        },
-    )
-    register_linter(
-        "jscs",
-        {
-            "install": [["npm", "install", "jscs"]],
-            "help": [os.path.normpath("./node_modules/.bin/jscs"), "-h"],
-            "run": [
-                os.path.normpath("./node_modules/.bin/jscs"),
-                ".",
-                "-r",
-                "json",
-                "-m",
-                "-1",
-                "-v",
-            ],
-            "rundefault": [
-                os.path.normpath("./node_modules/.bin/jscs"),
-                ".",
-                "-r",
-                "json",
-                "-m",
-                "-1",
-                "-v",
-                "-c",
-                "{config_dir}/.jscsrc",
-            ],
-            "dotfiles": [".jscsrc", ".jscs.json"],
-            "parser": parsers.JSCSParser,
-            "language": "javascript",
-            "autorun": False,
-            "run_per_file": False,
-        },
-    )
-    register_linter(
-        "jshint",
-        {
-            "install": [["npm", "install", "jshint"]],
-            "help": [os.path.normpath("./node_modules/.bin/jshint"), "-h"],
-            "run": [
-                os.path.normpath("./node_modules/.bin/jshint"),
-                ".",
-                "--reporter",
-                "checkstyle",
-            ],
-            "rundefault": [
-                os.path.normpath("./node_modules/.bin/jshint"),
-                ".",
-                "--reporter",
-                "checkstyle",
-                "-c",
-                "{config_dir}/.jshintrc",
-            ],
-            "dotfiles": [".jshintrc"],
-            "parser": parsers.JSHintParser,
-            "language": "javascript",
-            "autorun": False,
-            "run_per_file": False,
-        },
-    )
-    register_linter(
-        "jsonlint",
-        {
-            "install": [["npm", "install", "jsonlint"]],
-            "help": [os.path.normpath("./node_modules/.bin/jsonlint"), "-h"],
-            "run": [os.path.normpath("./node_modules/.bin/jsonlint"), "-c", "-q"],
-            "rundefault": [
-                os.path.normpath("./node_modules/.bin/jsonlint"),
-                "-c",
-                "-q",
-            ],
-            "dotfiles": [],
-            "parser": parsers.JSONLintParser,
-            "language": "json",
-            "autorun": True,
-            "run_per_file": True,
-        },
-    )
-    register_linter(
-        "markdownlint-cli",
-        {
-            "install": [["npm", "install", "markdownlint-cli"]],
-            "help": [os.path.normpath("./node_modules/.bin/markdownlint"), "-h"],
-            "run": [os.path.normpath("./node_modules/.bin/markdownlint"), "."],
-            "rundefault": [
-                os.path.normpath("./node_modules/.bin/markdownlint"),
-                ".",
-                "-c",
-                "{config_dir}/.markdownlintrc",
-            ],
-            "dotfiles": [".markdownlintrc", ".markdownlint.json"],
-            "parser": parsers.MarkdownLintParser,
-            "language": "markdown",
-            "autorun": True,
-            "run_per_file": False,
-        },
-    )
-    register_linter(
-        "megacheck",
-        {
-            "install": [["go", "get", "-u", "honnef.co/go/tools/cmd/megacheck"]],
-            "help": ["megacheck", "--help"],
-            "run": ["megacheck", "-f", "json", "./..."],
-            "rundefault": ["megacheck", "-f", "json", "./..."],
-            "dotfiles": [],
-            "parser": parsers.MegacheckParser,
-            "language": "go",
-            "autorun": False,
-            "run_per_file": False,
-        },
-    )
-    register_linter(
-        "pmd",
-        {
-            "install": [],
-            "help": [
-                vendored_path(os.path.join("pmd", "pmd-bin-6.3.0", "bin", "run.sh")),
-                "-help",
-            ],
-            "run": [
-                vendored_path(os.path.join("pmd", "pmd-bin-6.3.0", "bin", "run.sh")),
-                "pmd",
-                "-d",
-                ".",
-                "-R",
-                "java-basic",
-                "-f",
-                "emacs",
-            ],
-            "rundefault": [
-                vendored_path(os.path.join("pmd", "pmd-bin-6.3.0", "bin", "run.sh")),
-                "pmd",
-                "-d",
-                ".",
-                "-R",
-                "java-basic",
-                "-f",
-                "emacs",
-            ],
-            "dotfiles": [],
-            "parser": parsers.PMDParser,
-            "language": "java",
-            "autorun": True,
-            "run_per_file": False,
-        },
-    )
-    register_linter(
-        "proselint",
-        {
-            "install": [[sys.executable, "-m", "pip", "install", "-U", "proselint"]],
-            "help": ["proselint", "-h"],
-            "run": ["proselint"],
-            "rundefault": ["proselint"],
-            "dotfiles": [],
-            "parser": parsers.ProselintParser,
-            "language": "text",
-            "autorun": True,
-            "run_per_file": True,
-            "concurrency": 1,
-        },
-    )
-    register_linter(
-        "prospector",
-        {
-            "install": [
-                [
-                    sys.executable,
-                    "-m",
-                    "pip",
-                    "install",
-                    "-U",
-                    "prospector[with_everything]",
-                ],
-                [sys.executable, "-m", "pip", "install", "-U", "prospector"],
-            ],
-            "help": ["prospector", "-h"],
-            "run": ["prospector", "--zero-exit", "-o", "json"],
-            "rundefault": [
-                "prospector",
-                "--zero-exit",
-                "-o",
-                "json",
-                "-P",
-                "{config_dir}/.prospector.yaml",
-            ],
-            "dotfiles": [".prospector.yaml"],
-            "parser": parsers.ProspectorParser,
-            "language": "python",
-            "autorun": True,
-            "run_per_file": False,
-        },
-    )
-    register_linter(
-        "robotframework-lint",
-        {
-            "install": [
-                [sys.executable, "-m", "pip", "install", "-U", "robotframework-lint"]
-            ],
-            "help": ["rflint", "--help"],
-            "run": ["rflint"],
-            "rundefault": ["rflint", "-A", "{config_dir}/.rflint"],
-            "dotfiles": [".rflint"],
-            "parser": parsers.RobotFrameworkLintParser,
-            "language": "robotframework",
-            "autorun": True,
-            "run_per_file": True,
-        },
-    )
-    register_linter(
-        "restructuredtext_lint",
-        {
-            "install": [
-                [sys.executable, "-m", "pip", "install", "-U", "restructuredtext_lint"]
-            ],
-            "help": ["rst-lint", "-h"],
-            "run": ["rst-lint", "--format", "json", "--encoding", "utf-8"],
-            "rundefault": ["rst-lint", "--format", "json", "--encoding", "utf-8"],
-            "dotfiles": [],
-            "parser": parsers.RSTLintParser,
-            "language": "rst",
-            "autorun": True,
-            "run_per_file": True,
-        },
-    )
-    register_linter(
-        "shellcheck",
-        {
-            "install": [
-                ["cabal", "update"],
-                ["cabal", "install", "shellcheck"],
-                ["apt-get", "install", "shellcheck"],
-                ["dnf", "install", "shellcheck"],
-                ["brew", "install", "shellcheck"],
-                ["port", "install", "shellcheck"],
-                ["zypper", "in", "ShellCheck"],
-            ],
-            "help": ["shellcheck", "-V"],
-            "run": ["shellcheck", "-f", "json"],
-            "rundefault": ["shellcheck", "-f", "json"],
-            "dotfiles": [],
-            "parser": parsers.ShellcheckParser,
-            "language": "shell",
-            "autorun": True,
-            "run_per_file": True,
-        },
-    )
-    register_linter(
-        "spotbugs-maven-plugin",
-        {
-            "install": [
-                ["mvn", "clean", "install", "-U"],
-                [
-                    "mvn",
-                    "dependency:get",
-                    "-Dartifact=com.github.spotbugs:spotbugs-maven-plugin:3.1.3",
-                ],
-            ],
-            "help": ["mvn", "com.github.spotbugs:spotbugs-maven-plugin:3.1.3:help"],
-            "run": [
-                "mvn",
-                "-Dspotbugs.failOnError=false",
-                "com.github.spotbugs:spotbugs-maven-plugin:3.1.3:check",
-            ],
-            "rundefault": [
-                "mvn",
-                "-Dspotbugs.failOnError=false",
-                "com.github.spotbugs:spotbugs-maven-plugin:3.1.3:check",
-            ],
-            "dotfiles": [],
-            "parser": parsers.SpotbugsMavenParser,
-            "language": "java",
-            "autorun": True,
-            "run_per_file": False,
-        },
-    )
-    register_linter(
-        "stylint",
-        {
-            "install": [["npm", "install", "stylint"]],
-            "help": [os.path.normpath("./node_modules/.bin/stylint"), "-h"],
-            "run": [os.path.normpath("./node_modules/.bin/stylint")],
-            "rundefault": [
-                os.path.normpath("./node_modules/.bin/stylint"),
-                "-c",
-                "{config_dir}/.stylintrc",
-            ],
-            "dotfiles": [".stylintrc"],
-            "parser": parsers.StylintParser,
-            "language": "stylus",
-            "autorun": True,
-            "run_per_file": False,
-        },
-    )
-    register_linter(
-        "yamllint",
-        {
-            "install": [[sys.executable, "-m", "pip", "install", "yamllint"]],
-            "help": ["yamllint", "-h"],
-            "run": ["yamllint", "-f", "parsable", "."],
-            "rundefault": [
-                "yamllint",
-                "-c",
-                "{config_dir}/.yamllint",
-                "-f",
-                "parsable",
-                ".",
-            ],
-            "dotfiles": [".yamllint"],
-            "parser": parsers.YAMLLintParser,
-            "language": "yaml",
-            "autorun": True,
-            "run_per_file": False,
-        },
-    )
-
-
 def init():
     register_patterns()
-    register_linters()
 
 
 # TODO: can we not do this at "compile" or "import" time?
