+++ inlineplz/linters/bandit.py
@@ -2,11 +2,36 @@
 from __future__ import absolute_import
 from __future__ import unicode_literals
 
+import sys
 import dirtyjson as json
 
 from inlineplz.parsers.base import ParserBase
+from inlineplz.decorators import linter
 
 
+@linter(
+    {
+        "name": "bandit",
+        "install": [[sys.executable, "-m", "pip", "install", "-U", "bandit"]],
+        "help": ["bandit", "-h"],
+        "run": ["bandit", "-f", "json", "-iii", "-ll", "-r", "."],
+        "rundefault": [
+            "bandit",
+            "-f",
+            "json",
+            "-iii",
+            "-ll",
+            "-r",
+            ".",
+            "-c",
+            "{config_dir}/bandit.yaml",
+        ],
+        "dotfiles": ["bandit.yaml"],
+        "language": "python",
+        "autorun": True,
+        "run_per_file": False,
+    }
+)
 class BanditParser(ParserBase):
     """Parse json bandit output."""
