+++ inlineplz/linters/dockerfilelint.py
@@ -2,11 +2,29 @@
 from __future__ import absolute_import
 from __future__ import unicode_literals
 
+import os.path
 import dirtyjson as json
-
 from inlineplz.parsers.base import ParserBase
+from inlineplz.decorators import linter
 
 
+@linter(
+    {
+        "name": "dockerfile_lint",
+        "install": [["npm", "install", "dockerfile_lint"]],
+        "help": [os.path.normpath("./node_modules/.bin/dockerfile_lint"), "-h"],
+        "run": [os.path.normpath("./node_modules/.bin/dockerfile_lint"), "-j", "-f"],
+        "rundefault": [
+            os.path.normpath("./node_modules/.bin/dockerfile_lint"),
+            "-j",
+            "-f",
+        ],
+        "dotfiles": [],
+        "language": "docker",
+        "autorun": True,
+        "run_per_file": True,
+    }
+)
 class DockerfileLintParser(ParserBase):
     """Parse json dockerfile_lint output."""
