+++ inlineplz/linters/htmlhint.py
@@ -2,11 +2,30 @@
 from __future__ import absolute_import
 from __future__ import unicode_literals
 
+import os.path
 import dirtyjson as json
 
 from inlineplz.parsers.base import ParserBase
+from inlineplz.decorators import linter
 
 
+@linter(
+    {
+        "name": "htmlhint",
+        "install": [["npm", "install", "htmlhint"]],
+        "help": [os.path.normpath("./node_modules/.bin/htmlhint"), "-h"],
+        "run": [os.path.normpath("./node_modules/.bin/htmlhint"), "--format=json"],
+        "rundefault": [
+            os.path.normpath("./node_modules/.bin/htmlhint"),
+            "--format=json",
+            "--config={config_dir}/.htmlhintrc",
+        ],
+        "dotfiles": [".htmlhintrc"],
+        "language": "html",
+        "autorun": True,
+        "run_per_file": False,
+    }
+)
 class HTMLHintParser(ParserBase):
     """Parse json htmlhint output."""
