+++ inlineplz/linters/jshint.py
@@ -2,11 +2,38 @@
 from __future__ import absolute_import
 from __future__ import unicode_literals
 
+import os.path
 import xmltodict
 
 from inlineplz.parsers.base import ParserBase
+from inlineplz.decorators import linter
 
 
+@linter(
+    {
+        "name": "jshint",
+        "install": [["npm", "install", "jshint"]],
+        "help": [os.path.normpath("./node_modules/.bin/jshint"), "-h"],
+        "run": [
+            os.path.normpath("./node_modules/.bin/jshint"),
+            ".",
+            "--reporter",
+            "checkstyle",
+        ],
+        "rundefault": [
+            os.path.normpath("./node_modules/.bin/jshint"),
+            ".",
+            "--reporter",
+            "checkstyle",
+            "-c",
+            "{config_dir}/.jshintrc",
+        ],
+        "dotfiles": [".jshintrc"],
+        "language": "javascript",
+        "autorun": False,
+        "run_per_file": False,
+    }
+)
 class JSHintParser(ParserBase):
     """Parse json jshint output."""
