+++ inlineplz/linters/jsonlint.py
@@ -2,9 +2,25 @@
 from __future__ import absolute_import
 from __future__ import unicode_literals
 
+import os.path
+
 from inlineplz.parsers.base import ParserBase
+from inlineplz.decorators import linter
 
 
+@linter(
+    {
+        "name": "jsonlint",
+        "install": [["npm", "install", "jsonlint"]],
+        "help": [os.path.normpath("./node_modules/.bin/jsonlint"), "-h"],
+        "run": [os.path.normpath("./node_modules/.bin/jsonlint"), "-c", "-q"],
+        "rundefault": [os.path.normpath("./node_modules/.bin/jsonlint"), "-c", "-q"],
+        "dotfiles": [],
+        "language": "json",
+        "autorun": True,
+        "run_per_file": True,
+    }
+)
 class JSONLintParser(ParserBase):
     """Parse jsonlint output."""
