+++ inlineplz/linters/megacheck.py
@@ -5,8 +5,22 @@ from __future__ import unicode_literals
 import dirtyjson as json
 
 from inlineplz.parsers.base import ParserBase
+from inlineplz.decorators import linter
 
 
+@linter(
+    {
+        "name": "megacheck",
+        "install": [["go", "get", "-u", "honnef.co/go/tools/cmd/megacheck"]],
+        "help": ["megacheck", "--help"],
+        "run": ["megacheck", "-f", "json", "./..."],
+        "rundefault": ["megacheck", "-f", "json", "./..."],
+        "dotfiles": [],
+        "language": "go",
+        "autorun": False,
+        "run_per_file": False,
+    }
+)
 class MegacheckParser(ParserBase):
     """Parse json megacheck output."""
