+++ inlineplz/linters/shellcheck.py
@@ -5,8 +5,30 @@ from __future__ import unicode_literals
 import dirtyjson as json
 
 from inlineplz.parsers.base import ParserBase
+from inlineplz.decorators import linter
 
 
+@linter(
+    {
+        "name": "shellcheck",
+        "install": [
+            ["cabal", "update"],
+            ["cabal", "install", "shellcheck"],
+            ["apt-get", "install", "shellcheck"],
+            ["dnf", "install", "shellcheck"],
+            ["brew", "install", "shellcheck"],
+            ["port", "install", "shellcheck"],
+            ["zypper", "in", "ShellCheck"],
+        ],
+        "help": ["shellcheck", "-V"],
+        "run": ["shellcheck", "-f", "json"],
+        "rundefault": ["shellcheck", "-f", "json"],
+        "dotfiles": [],
+        "language": "shell",
+        "autorun": True,
+        "run_per_file": True,
+    }
+)
 class ShellcheckParser(ParserBase):
     """Parse json shellcheck output."""
