+++ inlineplz/linters/stylint.py
@@ -1,10 +1,29 @@
 # -*- coding: utf-8 -*-
 from __future__ import absolute_import
 from __future__ import unicode_literals
+import os.path
 
 from inlineplz.parsers.base import ParserBase
+from inlineplz.decorators import linter
 
 
+@linter(
+    {
+        "name": "stylint",
+        "install": [["npm", "install", "stylint"]],
+        "help": [os.path.normpath("./node_modules/.bin/stylint"), "-h"],
+        "run": [os.path.normpath("./node_modules/.bin/stylint")],
+        "rundefault": [
+            os.path.normpath("./node_modules/.bin/stylint"),
+            "-c",
+            "{config_dir}/.stylintrc",
+        ],
+        "dotfiles": [".stylintrc"],
+        "language": "stylus",
+        "autorun": True,
+        "run_per_file": False,
+    }
+)
 class StylintParser(ParserBase):
     """Parse stylint output."""
