+++ inlineplz/linters/tflint.py
@@ -7,22 +7,23 @@ from inlineplz.parsers.base import ParserBase
 import dirtyjson as json
 
 
-@linter()
+@linter(
+    {
+        "name": "tflint",
+        "language": "terraform",
+        "patterns": ["*.tf"],
+        "install": [["brew", "install", "tflint"]],
+        "help": ["tflint", "--help"],
+        "run": ["tflint", "--format=json"],
+        "rundefault": ["tflint", "--format=json"],
+        "dotfiles": [],
+        "autorun": True,
+        "run_per_file": False,
+    }
+)
 class TFLintParser(ParserBase):
     """Parse tflint output."""
 
-    name = "tflint"
-    language = "terraform"
-    patterns = ["*.tf"]
-    install = [["brew", "install", "tflint"]]
-    help = ["tflint", "--help"]
-    run = ["tflint", "--format=json"]
-    rundefault = ["tflint", "--format=json"]
-    dotfiles = []
-    language = "terraform"
-    autorun = True
-    run_per_file = False
-    patterns = ["*.tf"]
 
     def parse(self, lint_data):
         messages = set()
