+++ inlineplz/linters/yamllint.py
@@ -1,10 +1,32 @@
 # -*- coding: utf-8 -*-
 from __future__ import absolute_import
 from __future__ import unicode_literals
+import sys
 
 from inlineplz.parsers.base import ParserBase
+from inlineplz.decorators import linter
 
 
+@linter(
+    {
+        "name": "yamllint",
+        "install": [[sys.executable, "-m", "pip", "install", "yamllint"]],
+        "help": ["yamllint", "-h"],
+        "run": ["yamllint", "-f", "parsable", "."],
+        "rundefault": [
+            "yamllint",
+            "-c",
+            "{config_dir}/.yamllint",
+            "-f",
+            "parsable",
+            ".",
+        ],
+        "dotfiles": [".yamllint"],
+        "language": "yaml",
+        "autorun": True,
+        "run_per_file": False,
+    }
+)
 class YAMLLintParser(ParserBase):
     """Parse yaml-lint output."""
