+++ inlineplz/message.py
@@ -9,7 +9,7 @@ import os
 import traceback
 
 
-class Messages():
+class Messages:
     def __init__(self):
         self.messages = {}
 
@@ -45,7 +45,7 @@ class Messages():
         return self.messages.values()
 
 
-class Message():
+class Message:
     def __init__(self, path, line_number):
         self.path = os.path.relpath(path).replace("\\", "/")
         self.line_number = int(line_number)
