+++ inlineplz/parsers/__init__.py
@@ -3,29 +3,3 @@
 
 from __future__ import absolute_import
 from __future__ import unicode_literals
-
-from inlineplz.parsers.ansiblelint import AnsibleLintParser
-from inlineplz.parsers.bandit import BanditParser
-from inlineplz.parsers.codenarc import CodenarcParser
-from inlineplz.parsers.detectsecrets import DetectSecretsParser
-from inlineplz.parsers.dockerfilelint import DockerfileLintParser
-from inlineplz.parsers.eclint import ECLintParser
-from inlineplz.parsers.eslint import ESLintParser
-from inlineplz.parsers.gherkinlint import GherkinLintParser
-from inlineplz.parsers.gometalinter import GometalinterParser
-from inlineplz.parsers.htmlhint import HTMLHintParser
-from inlineplz.parsers.jscs import JSCSParser
-from inlineplz.parsers.jshint import JSHintParser
-from inlineplz.parsers.jsonlint import JSONLintParser
-from inlineplz.parsers.markdownlint import MarkdownLintParser
-from inlineplz.parsers.megacheck import MegacheckParser
-from inlineplz.parsers.pmd import PMDParser
-from inlineplz.parsers.proselint import ProselintParser
-from inlineplz.parsers.prospector import ProspectorParser
-from inlineplz.parsers.rstlint import RSTLintParser
-from inlineplz.parsers.rflint import RobotFrameworkLintParser
-from inlineplz.parsers.shellcheck import ShellcheckParser
-from inlineplz.parsers.spotbugsmaven import SpotbugsMavenParser
-from inlineplz.parsers.stylint import StylintParser
-from inlineplz.parsers.tflint import TFLintParser
-from inlineplz.parsers.yamllint import YAMLLintParser
