+++ tests/parsers/test_jscs.py
@@ -5,7 +5,7 @@ from __future__ import unicode_literals
 import codecs
 import os.path
 
-import inlineplz.parsers.jscs as jscs
+import inlineplz.linters.jscs as jscs
 
 jscs_path = os.path.join("tests", "testdata", "parsers", "jscs.txt")
