+++ tests/parsers/test_jsonlint.py
@@ -2,7 +2,7 @@
 from __future__ import absolute_import
 from __future__ import unicode_literals
 
-import inlineplz.parsers.jsonlint as jsonlint
+import inlineplz.linters.jsonlint as jsonlint
 
 
 def test_jsonlint():
