+++ tests/parsers/test_markdownlint.py
@@ -5,7 +5,7 @@ from __future__ import unicode_literals
 import codecs
 import os.path
 
-import inlineplz.parsers.markdownlint as markdownlint
+import inlineplz.linters.markdownlint as markdownlint
 
 markdownlint_path = os.path.join("tests", "testdata", "parsers", "markdownlint.txt")
