+++ tests/parsers/test_prospector.py
@@ -5,7 +5,7 @@ from __future__ import unicode_literals
 import codecs
 import os.path
 
-import inlineplz.parsers.prospector as prospector
+import inlineplz.linters.prospector as prospector
 
 prospector_path = os.path.join("tests", "testdata", "parsers", "prospector.txt")
