+++ tests/parsers/test_yamllint.py
@@ -5,7 +5,7 @@ from __future__ import unicode_literals
 import codecs
 import os
 
-import inlineplz.parsers.yamllint as yamllint
+import inlineplz.linters.yamllint as yamllint
 
 
 yamllint_path = os.path.join("tests", "testdata", "parsers", "yamllint.txt")
